+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
  - Add a footer
  - Add an explanation for why menu items have a disabled state ([#54](https://github.com/pwyf/iati-org-viz/issues/54))
  - Added the requisite codelists as static json files
+ - Add recipient country to document link listings ([#56](https://github.com/pwyf/iati-org-viz/issues/56))
 
 ### Changed
  - Prevent org ID badge from being orphaned